+++ src/applications/pensions/tests/validation.unit.spec.js
@@ -2,6 +2,7 @@ import { expect } from 'chai';
 import sinon from 'sinon';
 
 import {
+  isValidCurrency,
   validateAfterMarriageDate,
   validateServiceBirthDates,
 } from '../validation';
@@ -41,6 +42,32 @@ describe('Pension validation', () => {
       );
 
       expect(errors.activeServiceDateRange.from.addError.called).to.be.true;
+    });
+  });
+});
+
+describe('isValidCurrency', () => {
+  it('should return true for valid currency amount', () => {
+    const validCurrencyAmounts = ['1,000.00', '500', '0.99', '1,234,567.89'];
+
+    validCurrencyAmounts.forEach(amount => {
+      const result = isValidCurrency(amount);
+      expect(result).to.be.true;
+    });
+  });
+
+  it('should return false for invalid currency amount', () => {
+    const invalidCurrencyAmounts = [
+      '$1,000.00',
+      'abc',
+      '-$500',
+      '1.234',
+      '1.234.567',
+    ];
+
+    invalidCurrencyAmounts.forEach(amount => {
+      const result = isValidCurrency(amount);
+      expect(result).to.be.false;
     });
   });
 });
