+++ src/applications/pensions/validation.js
@@ -52,3 +52,9 @@ export function validateCentralMailPostalCode(errors, address) {
     );
   }
 }
+
+// Does not allow a dollar sign
+export const isValidCurrency = currencyAmount => {
+  const regex = /^(?!.*\$)(([1-9]\d{0,2}(,\d{3})*|\d+)(\.\d{1,2})?)?$/;
+  return regex.test(currencyAmount);
+};
